+++ sciris/scirisobjects.py
@@ -1,7 +1,7 @@
 """
 scirisobjects.py -- classes for Sciris objects which are generally managed
     
-Last update: 11/22/17 (gchadder3)
+Last update: 5/25/18 (gchadder3)
 """
 
 #
@@ -22,256 +22,244 @@ class ScirisObject(object):
     functionality.
     
     Methods:
-        __init__(theUID: UUID [None], theTypePrefix: str ['obj'], 
-            theFileSuffix: str ['.obj'], theInstanceLabel: str ['']): 
+        __init__(uid: UUID [None], type_prefix: str ['obj'], 
+            file_suffix: str ['.obj'], instance_label: str ['']): 
             void -- constructor
-        inDataStore(): bool -- is the object stored in the DataStore?
-        loadFromCopy(otherObject): void -- assuming otherObject is another 
+        in_data_store(): bool -- is the object stored in the DataStore?
+        load_from_copy(other_obj): void -- assuming other_obj is another 
             object of our type, copy its contents to us
-        addToDataStore(): void -- add ourselves to the Sciris DataStore 
+        add_to_data_store(): void -- add ourselves to the Sciris DataStore 
             (theDataStore object)
-        loadCopyFromDataStore(theUID: UUID [None]): ScirisObject -- return 
+        load_copy_from_data_store(uid: UUID [None]): ScirisObject -- return 
             the ScirisObject matching a UUID (usually ourselves in UUID is 
             None)
-        loadFromDataStore(self, theUID=None): void -- overwrite our current 
+        load_from_data_store(self, uid=None): void -- overwrite our current 
             state with what's in the DataStore (usually from our stored 
             object, but you can pass in another UUID also to copy from another 
             stored object)
-        updateDataStore(): void -- update the DataStore representation from 
+        update_data_store(): void -- update the DataStore representation from 
             our current state
-        deleteFromDataStore(): void -- delete ourselves from the DataStore        
+        delete_from_data_store(): void -- delete ourselves from the DataStore        
         show(): void -- print the contents of the object       
-        getUserFrontEndRepr(): dict -- get a JSON-friendly dictionary 
+        get_user_front_end_repr(): dict -- get a JSON-friendly dictionary 
             representation of the object state the front-end uses for non-
             admin purposes
-        getAdminFrontEndRepr(): dict -- get a JSON-friendly dictionary
+        get_admin_front_end_repr(): dict -- get a JSON-friendly dictionary
             representation of the object state the front-end uses for admin
             purposes
                     
     Attributes:
         uid (UUID) -- the unique ID for the user (uuid Python library-related)
-        typePrefix (str) -- a prefix that gets added to the UUID to give either 
+        type_prefix (str) -- a prefix that gets added to the UUID to give either 
             a file name or a Redis key
-        fileSuffix (str) -- a suffix that gets added to files
-        instanceLabel (str) -- a name of the object which should at least be 
-            unique across other handles of the save typePrefix
+        file_suffix (str) -- a suffix that gets added to files
+        instance_label (str) -- a name of the object which should at least be 
+            unique across other handles of the save type_prefix
             
     Usage:
-        >>> theObj = ScirisObject(uuid.UUID('12345678123456781234567812345678'))                      
+        >>> obj = ScirisObject(uuid.UUID('12345678123456781234567812345678'))                      
     """
     
-    def  __init__(self, theUID=None, theTypePrefix='obj', theFileSuffix='.obj', 
-        theInstanceLabel=''):        
-        # If a UUID was passed in...
-        if theUID is not None:
-            # Make sure the argument is a valid UUID, converting a hex text to a
-            # UUID object, if needed.        
-            validUID = getValidUUID(theUID) 
-            
-            # If a validUID was found, use it.
-            if validUID is not None:
-                self.uid = validUID
-            # Otherwise, generate a new random UUID using uuid4().
-            else:
-                self.uid = uuid.uuid4()
-        # Otherwise, generate a new random UUID using uuid4().
-        else:
-            self.uid = uuid.uuid4()
+    def  __init__(self, uid=None, type_prefix='obj', file_suffix='.obj', 
+        instance_label=''):       
+        # Get a valid UUID from what is passed in, or if None is passed in, 
+        # get a new ID.
+        self.uid = get_valid_uuid(uid, new_uuid_if_missing=True) 
             
         # Set the other variables that might be used with DataStore.
-        self.typePrefix = theTypePrefix
-        self.fileSuffix = theFileSuffix
-        self.instanceLabel = theInstanceLabel
+        self.type_prefix = type_prefix
+        self.file_suffix = file_suffix
+        self.instance_label = instance_label
         
-    def inDataStore(self):
-        return (ds.theDataStore.retrieve(self.uid) is not None)
+    def in_data_store(self):
+        return (ds.data_store.retrieve(self.uid) is not None)
     
-    def loadFromCopy(self, otherObject):
-        if type(otherObject) == type(self):
-            self.uid = otherObject.uid
-            self.typePrefix = otherObject.typePrefix
-            self.instanceLabel = otherObject.instanceLabel
+    def load_from_copy(self, other_obj):
+        if type(other_obj) == type(self):
+            self.uid = other_obj.uid
+            self.type_prefix = other_obj.type_prefix
+            self.instance_label = other_obj.instance_label
             
-    def addToDataStore(self):
+    def add_to_data_store(self):
         # Check to see if the object is already in the DataStore, and give an 
         # error if so.
-        if self.inDataStore():
+        if self.in_data_store():
             print 'Error: Object ''%s'' is already in DataStore.' % self.uid.hex
             return
         
         # Add our representation to the DataStore.
-        ds.theDataStore.add(self, self.uid, self.typePrefix, self.fileSuffix, 
-            self.instanceLabel)
+        ds.data_store.add(self, self.uid, self.type_prefix, self.file_suffix, 
+            self.instance_label)
         
-    def loadCopyFromDataStore(self, theUID=None):
-        if theUID is None:
-            if not self.inDataStore():
+    def load_copy_from_data_store(self, uid=None):
+        if uid is None:
+            if not self.in_data_store():
                 print 'Error: Object ''%s'' is not in DataStore.' % self.uid.hex
                 return None
             else:
-                return ds.theDataStore.retrieve(self.uid)
+                return ds.data_store.retrieve(self.uid)
         else:
-            return ds.theDataStore.retrieve(theUID)
+            return ds.data_store.retrieve(uid)
             
-    def loadFromDataStore(self, theUID=None):
+    def load_from_data_store(self, uid=None):
         # Get a copy from the DataStore.
-        copyFromStore = self.loadCopyFromDataStore(theUID)
+        copy_from_store = self.load_copy_from_data_store(uid)
         
         # Copy the internal information over from the copy to ourselves.
-        self.loadFromCopy(copyFromStore)
+        self.load_from_copy(copy_from_store)
         
-    def updateDataStore(self):
+    def update_data_store(self):
         # Give an error if the object is not in the DataStore.
-        if not self.inDataStore():
+        if not self.in_data_store():
             print 'Error: Object ''%s'' is not in DataStore.' % self.uid.hex
             return
         
         # Update our DataStore representation with our current state. 
-        ds.theDataStore.update(self.uid, self) 
+        ds.data_store.update(self.uid, self) 
         
-    def deleteFromDataStore(self):
+    def delete_from_data_store(self):
         # Give an error if the object is not in the DataStore.
-        if not self.inDataStore():
+        if not self.in_data_store():
             print 'Error: Object ''%s'' is not in DataStore.' % self.uid.hex
             return
         
         # Delete ourselves from the DataStore.
-        ds.theDataStore.delete(self.uid)
+        ds.data_store.delete(self.uid)
     
     def show(self):
         print '--------------------------------------------'
         print 'UUID: %s' % self.uid.hex
-        print 'Type Prefix: %s' % self.typePrefix
-        print 'File Suffix: %s' % self.fileSuffix
-        print 'Instance Label: %s' % self.instanceLabel
-        inDataStore = self.inDataStore()
-        if inDataStore:
+        print 'Type Prefix: %s' % self.type_prefix
+        print 'File Suffix: %s' % self.file_suffix
+        print 'Instance Label: %s' % self.instance_label
+        in_data_store = self.in_data_store()
+        if in_data_store:
             print 'In DataStore?: Yes'
         else:
             print 'In DataStore?: No'
         #print '--------------------------------------------'
         
-    def getUserFrontEndRepr(self):
-        objInfo = {
+    def get_user_front_end_repr(self):
+        obj_info = {
             'scirisobject': {
-                'instancelabel': self.instanceLabel                
+                'instance_label': self.instance_label                
             }
         }
-        return objInfo
+        return obj_info
     
-    def getAdminFrontEndRepr(self):
-        objInfo = {
+    def get_admin_front_end_repr(self):
+        obj_info = {
             'scirisobject': {
                 'UID': self.uid.hex, 
-                'typeprefix': self.typePrefix, 
-                'filesuffix': self.fileSuffix, 
-                'instancelabel': self.instanceLabel                
+                'type_prefix': self.type_prefix, 
+                'file_suffix': self.file_suffix, 
+                'instance_label': self.instance_label                
             }
         }
-        return objInfo 
+        return obj_info 
           
 class ScirisCollection(ScirisObject):
     """
-    A collection of ScirisObjects (stored in a Python dict theObjectDict).
+    A collection of ScirisObjects (stored in a Python dict obj_dict).
     
     Methods:
-        __init__(theUID: UUID [None], theTypePrefix: str ['collection'], 
-            theFileSuffix: str ['.scl'], theInstanceLabel: str ['']): 
+        __init__(uid: UUID [None], type_prefix: str ['collection'], 
+            file_suffix: str ['.scl'], instance_label: str ['']): 
             void -- constructor        
-        loadFromCopy(otherObject): void -- assuming otherObject is another 
+        load_from_copy(other_obj): void -- assuming other_obj is another 
             object of our type, copy its contents to us (calls the 
             ScirisObject superclass version of this method also)
-        getObjectByUID(theUID: UUID): ScirisObject -- get a ScirisObject out of 
+        get_object_by_uid(uid: UUID): ScirisObject -- get a ScirisObject out of 
             the collection by the UUID passed in
-        getAllObjects(): list of ScirisObjects -- get all of the ScirisObjects 
+        get_all_objects(): list of ScirisObjects -- get all of the ScirisObjects 
            and put them in a list
-        addObject(theObject: ScirisObject): void -- add a ScirisObject to the 
+        add_object(obj: ScirisObject): void -- add a ScirisObject to the 
             collection
-        updateObject(theObject: ScirisObject): void -- update a ScirisObject 
+        update_object(obj: ScirisObject): void -- update a ScirisObject 
             passed in to the collection
-        deleteObjectByUID(theUID: UUID): void -- delete the ScirisObject 
+        delete_object_by_uid(uid: UUID): void -- delete the ScirisObject 
             indexed by the UUID from the collection
-        deleteAllObjects(): void -- delete all ScirisObjects from the 
+        delete_all_objects(): void -- delete all ScirisObjects from the 
             collection
         show(): void -- print the contents of the collection, including the 
             object information as well as the objects
                     
     Attributes:
-        theObjectDict (dict) -- the Python dictionary holding the ScirisObjects
+        obj_dict (dict) -- the Python dictionary holding the ScirisObjects
         
     Usage:
-        >>> theObjs = ScirisCollection(uuid.UUID('12345678123456781234567812345678'))                      
+        >>> objs = ScirisCollection(uuid.UUID('12345678123456781234567812345678'))                      
     """
     
-    def __init__(self, theUID, theTypePrefix='collection', 
-        theFileSuffix='.scl', theInstanceLabel=''):
+    def __init__(self, uid, type_prefix='collection', 
+        file_suffix='.scl', instance_label=''):
         # Set superclass parameters.
-        super(ScirisCollection, self).__init__(theUID, theTypePrefix, 
-             theFileSuffix, theInstanceLabel)
+        super(ScirisCollection, self).__init__(uid, type_prefix, 
+             file_suffix, instance_label)
         
         # Create a Python dict to hold the ScirisObjects.
-        self.theObjectDict = {}
+        self.obj_dict = {}
         
-    def loadFromCopy(self, otherObject):
-        if type(otherObject) == type(self):
+    def load_from_copy(self, other_obj):
+        if type(other_obj) == type(self):
             # Do the superclass copying.
-            super(ScirisCollection, self).loadFromCopy(otherObject)
+            super(ScirisCollection, self).load_from_copy(other_obj)
             
-            self.theObjectDict = otherObject.theObjectDict
+            self.obj_dict = other_obj.obj_dict
             
-    def getObjectByUID(self, theUID):
+    def get_object_by_uid(self, uid):
         # Make sure the argument is a valid UUID, converting a hex text to a
         # UUID object, if needed.
-        validUID = getValidUUID(theUID)
+        valid_uid = get_valid_uuid(uid)
         
         # If we have a valid UUID, return the matching ScirisObject (if any); 
         # otherwise, return None.
-        if validUID is not None:
-            return self.theObjectDict.get(validUID, None)
+        if valid_uid is not None:
+            return self.obj_dict.get(valid_uid, None)
         else:
             return None
         
-    def getAllObjects(self):
-        return [self.theObjectDict[theKey] for theKey in self.theObjectDict]
+    def get_all_objects(self):
+        return [self.obj_dict[key] for key in self.obj_dict]
     
-    def addObject(self, theObject):
+    def add_object(self, obj):
         # Add the object to the hash table, keyed by the UID.
-        self.theObjectDict[theObject.uid] = theObject
+        self.obj_dict[obj.uid] = obj
         
         # Update our DataStore representation if we are there. 
-        if self.inDataStore():
-            self.updateDataStore()
+        if self.in_data_store():
+            self.update_data_store()
             
-    def updateObject(self, theObject):
-        # Do the same behavior as addObject().
-        self.addObject(theObject)
+    def update_object(self, obj):
+        # Do the same behavior as add_object().
+        self.add_object(obj)
             
-    def deleteObjectByUID(self, theUID):
+    def delete_object_by_uid(self, uid):
         # Make sure the argument is a valid UUID, converting a hex text to a
         # UUID object, if needed.        
-        validUID = getValidUUID(theUID)
+        valid_uid = get_valid_uuid(uid)
         
         # If we have a valid UUID...
-        if validUID is not None:
+        if valid_uid is not None:
             # Get the object pointed to by the UID.
-            theObject = self.theObjectDict[validUID]
+            obj = self.obj_dict[valid_uid]
             
             # If a match is found...
-            if theObject is not None:
-                # Remove entries from theObjectDict.
-                del self.theObjectDict[validUID]
+            if obj is not None:
+                # Remove entries from obj_dict.
+                del self.obj_dict[valid_uid]
                 
                 # Update our DataStore representation if we are there. 
-                if self.inDataStore():
-                    self.updateDataStore()
+                if self.in_data_store():
+                    self.update_data_store()
     
-    def deleteAllObjects(self):
+    def delete_all_objects(self):
         # Reset the Python dicts.
-        self.theObjectDict = {}
+        self.obj_dict = {}
         
         # Update our DataStore representation if we are there. 
-        if self.inDataStore():
-            self.updateDataStore()  
+        if self.in_data_store():
+            self.update_data_store()  
         
     def show(self):
         # Show superclass attributes.
@@ -282,15 +270,15 @@ class ScirisCollection(ScirisObject):
         print '---------------------'
         
         # For each key in the dictionary...
-        for theKey in self.theObjectDict:
+        for key in self.obj_dict:
             # Get the object pointed to.
-            theObject = self.theObjectDict[theKey]
+            obj = self.obj_dict[key]
             
             # Separator line.
             #print '--------------------------------------------'
             
             # Show the handle contents.
-            theObject.show()
+            obj.show()
             
         # Separator line.
         print '--------------------------------------------'
@@ -299,19 +287,24 @@ class ScirisCollection(ScirisObject):
 # Other utility functions
 #
 
-def getValidUUID(uidParam):
+def get_valid_uuid(uid_param, new_uuid_if_missing=False):
     # Get the type of the parameter passed in.
-    paramType = type(uidParam)
+    param_type = type(uid_param)
     
     # Return what was passed in if it is already the right type.
-    if paramType == uuid.UUID:
-        return uidParam
+    if param_type == uuid.UUID:
+        return uid_param
     
     # Try to do the conversion and if it fails, set the conversion to None.
     try:
-        convertParam = uuid.UUID(uidParam)
+        convert_param = uuid.UUID(uid_param)
     except:
-        convertParam = None
-    
+        convert_param = None
+        
+    # If we don't have a valid UUID and we are supposed to generate a new 
+    # one if it's missing, do so using uuid4() which generates a random ID.
+    if convert_param is None and new_uuid_if_missing:
+        convert_param = uuid.uuid4()
+        
     # Return the converted value.
-    return convertParam 
+    return convert_param
