+++ sciris/test/test_scirisobjects.py
