+++ sciris/version.py
@@ -1,2 +1,2 @@
-version = '0.4.1'
-versiondate = '2018-05-18'
+version = '0.5'
+versiondate = '2018-05-31'
