--- sciris2gc/version.py
@@ -1,2 +1,0 @@
-version = '0.5'
-versiondate = '2018-05-31'
