+++ sciris_legacy/version.py
@@ -1,0 +1,2 @@
+version = '0.4.1'
+versiondate = '2018-05-18'
